+++ lib/netutil/conn_pool_test.go
@@ -7,6 +7,7 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"sync"
+	"sync/atomic"
 	"testing"
 	"time"
 
@@ -49,7 +50,7 @@ func testConnPoolStartStop(t *testing.T, name string, ms *metrics.Set) {
 	var cps []*ConnPool
 	for i := range 5 {
 		addr := fmt.Sprintf("host-%d", i)
-		cp := NewConnPool(ms, name, addr, handshake.VMSelectClient, compressLevel, dialTimeout, 0)
+		cp := NewConnPool(ms, name, addr, handshake.VMSelectClient, compressLevel, dialTimeout, 0, mockHealthCheck)
 		cps = append(cps, cp)
 	}
 	for _, cp := range cps {
@@ -76,7 +77,7 @@ func testConnPoolStartStop(t *testing.T, name string, ms *metrics.Set) {
 func TestGetPutDialConnectionPool(t *testing.T) {
 	mockSvr := newMockServer()
 	addr, _ := url.Parse(mockSvr.URL)
-	cp := NewConnPool(metrics.NewSet(), "test-pool", addr.Host, mockHandshake, 1, 5*time.Second, 0)
+	cp := NewConnPool(metrics.NewSet(), "test-pool", addr.Host, mockHandshake, 1, 5*time.Second, 0, mockHealthCheck)
 
 	concurrency := 256
 	connChan := make(chan *handshake.BufferedConn, concurrency)
@@ -105,6 +106,76 @@ func TestGetPutDialConnectionPool(t *testing.T) {
 	wg.Wait()
 }
 
+func TestGetWithHealthCheckConnectionPool(t *testing.T) {
+	mockSvr := newMockServer()
+	addr, _ := url.Parse(mockSvr.URL)
+	healthCheckCounter := atomic.Int32{}
+	handshakeCounter := atomic.Int32{}
+
+	mockHealthCheckWithCounter := func(_ *handshake.BufferedConn) error {
+		healthCheckCounter.Add(1)
+		return nil
+	}
+	mockFailHealthCheckWithCounter := func(_ *handshake.BufferedConn) error {
+		healthCheckCounter.Add(1)
+		return fmt.Errorf("fail health check")
+	}
+	mockHandshakeWithCounter := func(c net.Conn, n int) (*handshake.BufferedConn, error) {
+		handshakeCounter.Add(1)
+		return mockHandshake(c, n)
+	}
+	cp := NewConnPool(metrics.NewSet(), "test-pool", addr.Host, mockHandshakeWithCounter, 1, 5*time.Second, 0, mockHealthCheckWithCounter)
+	cp.connKeepAliveInterval = 3
+	// generate new connections and put it in the pool.
+	conn, err := cp.Get()
+	if err != nil {
+		t.Errorf("get conn from connection pool err:%v", err)
+		panic(err)
+	}
+	if handshakeCounter.Load() != 1 {
+		t.Errorf("unexpected handshake counter value: %d", handshakeCounter.Load())
+	}
+	cp.Put(conn)
+
+	// get an existing active connection
+	conn, err = cp.Get()
+	if conn == nil || err != nil {
+		t.Errorf("failed to get connection from pool")
+	}
+	if conn != nil {
+		cp.Put(conn)
+	}
+	if healthCheckCounter.Load() != 0 {
+		t.Errorf("unexpected health check counter value: %d", healthCheckCounter.Load())
+	}
+	if handshakeCounter.Load() != 1 {
+		t.Errorf("unexpected handshake counter value: %d", handshakeCounter.Load())
+	}
+
+	// wait for connection to be in-active
+	// and then get an existing in-active connection
+	time.Sleep(4 * time.Second)
+	conn, err = cp.Get()
+	if conn == nil || err != nil {
+		t.Errorf("failed to get connection from pool")
+	}
+	if healthCheckCounter.Load() != 1 {
+		t.Errorf("unexpected health check counter value: %d", healthCheckCounter.Load())
+	}
+
+	// health check fail, connection will dial a new connection to return.
+	cp.healthCheckFunc = mockFailHealthCheckWithCounter
+
+	time.Sleep(4 * time.Second)
+	conn, err = cp.Get()
+	if conn == nil || err != nil {
+		t.Errorf("failed to get connection from pool")
+	}
+	if handshakeCounter.Load() != 2 {
+		t.Errorf("unexpected handshake counter value: %d", handshakeCounter.Load())
+	}
+}
+
 // mockServer does nothing. It only acts as a tcp server for connection test.
 type mockServer struct {
 	*httptest.Server
@@ -123,4 +194,8 @@ func mockHandshake(c net.Conn, _ int) (*handshake.BufferedConn, error) {
 		Conn: c,
 	}
 	return bc, nil
+}
+
+func mockHealthCheck(_ *handshake.BufferedConn) error {
+	return nil
 }